+++ include/daScript/ast/ast_policy_types.h
@@ -9,7 +9,7 @@ namespace  das {
             noSideEffects = true;
             this->result = makeType<RetT>(lib);
             vector<TypeDeclPtr> args = { makeType<Args>(lib)... };
-            for ( int argi=0; argi != args.size(); ++argi ) {
+            for ( size_t argi=0; argi != args.size(); ++argi ) {
                 auto arg = make_shared<Variable>();
                 arg->name = "arg" + std::to_string(argi);
                 arg->type = args[argi];
